requirements.txt
python-dotenv 
typer
requests
httpx
pandas
firecrawl
scrapingbee
apify-client
scrapy
crawl4ai
playwright
selenium
exa-py
tavily-python
plotly
nbformat